backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==23.0.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.2
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.12
openai>=1.60.0
httpx==0.27.0
Pillow==10.4.0
cryptography>=42.0.0
slowapi==0.1.9
aiosmtplib==3.0.2
sentry-sdk[fastapi]>=2.0.0
PyJWT[crypto]>=2.8.0

# ─── Dev / Test ───
pytest>=8.0
pytest-asyncio>=0.23
aiosqlite>=0.20.0
